Fix bug in Element::fromSymbol()
This fixes a small bug in the fromSymbol() method with the
symbol and length arguments. Also adds a test to ensure the
bug doesn't occur again.

+++ src/chemkit/element.h
@@ -82,7 +82,7 @@ public:
     static Element fromName(const char *name);
     static Element fromSymbol(const std::string &symbol);
     static Element fromSymbol(const char *symbol);
-    static Element fromSymbol(const char *symbol, int length);
+    static Element fromSymbol(const char *symbol, size_t length);
     static Element fromSymbol(char symbol);
     static bool isValidAtomicNumber(AtomicNumberType atomicNumber);
     static bool isValidSymbol(const std::string &symbol);
